+++ junit-platform-launcher/src/main/java/org/junit/platform/launcher/LauncherInterceptor.java
@@ -11,9 +11,6 @@
 package org.junit.platform.launcher;
 
 import static org.apiguardian.api.API.Status.EXPERIMENTAL;
-import static org.apiguardian.api.API.Status.INTERNAL;
-
-import java.util.List;
 
 import org.apiguardian.api.API;
 
@@ -58,18 +55,6 @@ import org.apiguardian.api.API;
 @API(status = EXPERIMENTAL, since = "1.10")
 public interface LauncherInterceptor {
 
-	LauncherInterceptor NOOP = new LauncherInterceptor() {
-		@Override
-		public <T> T intercept(Invocation<T> invocation) {
-			return invocation.proceed();
-		}
-
-		@Override
-		public void close() {
-			// do nothing
-		}
-	};
-
 	/**
 	 * Intercept the supplied invocation.
 	 *
@@ -97,28 +82,4 @@ public interface LauncherInterceptor {
 		T proceed();
 	}
 
-	@API(status = INTERNAL, since = "1.10")
-	static LauncherInterceptor composite(List<LauncherInterceptor> interceptors) {
-		if (interceptors.isEmpty()) {
-			return NOOP;
-		}
-		return interceptors.stream() //
-				.skip(1) //
-				.reduce(interceptors.get(0), (a, b) -> new LauncherInterceptor() {
-					@Override
-					public void close() {
-						try {
-							a.close();
-						}
-						finally {
-							b.close();
-						}
-					}
-
-					@Override
-					public <T> T intercept(Invocation<T> invocation) {
-						return a.intercept(() -> b.intercept(invocation));
-					}
-				});
-	}
 }
